+++ corehq/apps/linked_domain/exceptions.py
@@ -1,0 +1,11 @@
+
+class RemoteRequestError(Exception):
+    pass
+
+
+class RemoteAuthError(RemoteRequestError):
+    pass
+
+
+class ActionNotPermitted(RemoteRequestError):
+    pass
